cherrypick #933 #904 (#936)
* Check type of index (#933)

(cherry picked from commit ef5db1e1fffddb700da173dbf53b44a724661c8f)

* Set thread manager for thrift server (#897)

(cherry picked from commit 5d5e6f92c79bf7e0fb291300f97c60902cbe2092)

* Fix unstable test cases related to index (#904)

* Fix unstable test cases related to index

* Format

* Fix package shell script

Using ShellCheck to static check bash shell: https://github.com/koalaman/shellcheck

* Migrate lookup in stateless

(cherry picked from commit d965a3636507f751b7388ae17cf9f661c6ddb56a)

* #904

* update

+++ package/package.sh
@@ -55,43 +55,46 @@ do
 done
 
 # version is null, get from tag name
-[[ -z $version ]] && version=`git describe --exact-match --abbrev=0 --tags | sed 's/^v//'`
+[[ -z $version ]] && version=$(git describe --exact-match --abbrev=0 --tags | sed 's/^v//')
 # version is null, use UTC date as version
 [[ -z $version ]] && version=$(date -u +%Y.%m.%d)-nightly
 
 if [[ -z $version ]]; then
     echo "version is null, exit"
     echo ${usage}
-    exit -1
+    exit 1
 fi
 
 
 if [[ $strip_enable != TRUE ]] && [[ $strip_enable != FALSE ]]; then
     echo "strip enable is wrong, exit"
     echo ${usage}
-    exit -1
+    exit 1
 fi
 
 echo "current version is [ $version ], strip enable is [$strip_enable], enablesanitizer is [$enablesanitizer], static_sanitizer is [$static_sanitizer]"
 
-# args: <version>
-function build {
-    version=$1
-    san=$2
-    ssan=$3
-    build_type=$4
-    branch=$5
-    modules_dir=${project_dir}/modules
-    if [[ -d $build_dir ]]; then
-        rm -rf ${build_dir}/*
-    else
-        mkdir ${build_dir}
+function _build_storage {
+    if [[ ! -d ${storage_dir} && ! -L ${storage_dir} ]]; then
+        git clone --single-branch --branch ${branch} https://github.com/vesoft-inc/nebula-storage.git ${storage_dir}
     fi
 
-    if [[ -d $modules_dir ]]; then
-        rm -rf ${modules_dir}/*
-    else
-        mkdir ${modules_dir}
+    cmake -DCMAKE_BUILD_TYPE=${build_type} \
+          -DNEBULA_BUILD_VERSION=${version} \
+          -DENABLE_ASAN=${san} \
+          -DENABLE_UBSAN=${san} \
+          -DENABLE_STATIC_ASAN=${ssan} \
+          -DENABLE_STATIC_UBSAN=${ssan} \
+          -DCMAKE_INSTALL_PREFIX=/usr/local/nebula \
+          -DNEBULA_COMMON_REPO_TAG=${branch} \
+          -DENABLE_TESTING=OFF \
+          -DENABLE_PACK_ONE=${package_one} \
+          -S ${storage_dir} \
+          -B ${storage_build_dir}
+
+    if ! ( cmake --build ${storage_build_dir} -j ${jobs} ); then
+        echo ">>> build nebula storage failed <<<"
+        exit 1
     fi
 
     pushd ${build_dir}
@@ -110,9 +113,9 @@ function build {
           -DENABLE_PACK_ONE=${package_one} \
           $project_dir
 
-    if !( make -j$(nproc) ); then
-        echo ">>> build nebula failed <<<"
-        exit -1
+    if ! ( cmake --build ${build_dir} -j ${jobs} ); then
+        echo ">>> build nebula graph failed <<<"
+        exit 1
     fi
 
     popd
@@ -123,7 +126,7 @@ function package {
     # The package CMakeLists.txt in ${project_dir}/package/build
     package_dir=${build_dir}/package/
     if [[ -d $package_dir ]]; then
-        rm -rf ${package_dir}/*
+        rm -rf ${package_dir:?}/*
     else
         mkdir ${package_dir}
     fi
@@ -142,30 +145,30 @@ function package {
     sys_ver=""
     pType="RPM"
     if [[ -f "/etc/redhat-release" ]]; then
-        sys_name=`cat /etc/redhat-release | cut -d ' ' -f1`
+        sys_name=$(< /etc/redhat-release cut -d ' ' -f1)
         if [[ ${sys_name} == "CentOS" ]]; then
-            sys_ver=`cat /etc/redhat-release | tr -dc '0-9.' | cut -d \. -f1`
+            sys_ver=$(< /etc/redhat-release tr -dc '0-9.' | cut -d \. -f1)
             sys_ver=.el${sys_ver}.x86_64
         elif [[ ${sys_name} == "Fedora" ]]; then
-            sys_ver=`cat /etc/redhat-release | cut -d ' ' -f3`
+            sys_ver=$(< /etc/redhat-release cut -d ' ' -f3)
             sys_ver=.fc${sys_ver}.x86_64
         fi
         pType="RPM"
     elif [[ -f "/etc/lsb-release" ]]; then
-        sys_ver=`cat /etc/lsb-release | grep DISTRIB_RELEASE | cut -d "=" -f 2 | sed 's/\.//'`
+        sys_ver=$(< /etc/lsb-release grep DISTRIB_RELEASE | cut -d "=" -f 2 | sed 's/\.//')
         sys_ver=.ubuntu${sys_ver}.amd64
         pType="DEB"
     fi
 
-    if !( cpack -G ${pType} --verbose $args ); then
+    if ! ( cpack -G ${pType} --verbose $args ); then
         echo ">>> package nebula failed <<<"
-        exit -1
+        exit 1
     else
         # rename package file
-        pkg_names=`ls | grep nebula | grep ${version}`
+        pkg_names=$(ls ./*nebula*-${version}*)
         outputDir=$build_dir/cpack_output
         mkdir -p ${outputDir}
-        for pkg_name in ${pkg_names[@]};
+        for pkg_name in "${pkg_names[@]}";
         do
             new_pkg_name=${pkg_name/\-Linux/${sys_ver}}
             mv ${pkg_name} ${outputDir}/${new_pkg_name}
